Move around files with new seeding package
