+++ server/app/modules/DatabaseSeedModule.java
@@ -3,10 +3,10 @@ package modules;
 import akka.actor.ActorSystem;
 import com.google.inject.AbstractModule;
 import com.google.inject.Inject;
+import controllers.dev.seeding.DatabaseSeedTask;
 import java.time.Duration;
 import javax.inject.Provider;
 import scala.concurrent.ExecutionContext;
-import tasks.DatabaseSeedTask;
 
 /**
  * Binds the {@link DatabaseSeedScheduler} as an eager singleton, which causes it to run at server
